+++ commands/cloudapi-v6/k8s_cluster_test.go
@@ -31,6 +31,22 @@ var (
 					},
 				},
 				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPublicBoolVar,
+			},
+		},
+	}
+	clusterPrivateTestPost = resources.K8sClusterForPost{
+		KubernetesClusterForPost: ionoscloud.KubernetesClusterForPost{
+			Properties: &ionoscloud.KubernetesClusterPropertiesForPost{
+				Name:       &testClusterVar,
+				K8sVersion: &testClusterVar,
+				S3Buckets: &[]ionoscloud.S3Bucket{
+					{
+						Name: &testClusterVar,
+					},
+				},
+				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPrivateBoolVar,
 			},
 		},
 	}
@@ -77,6 +93,22 @@ var (
 					},
 				},
 				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPublicBoolVar,
+			},
+		},
+	}
+	clusterPrivateTest = resources.K8sCluster{
+		KubernetesCluster: ionoscloud.KubernetesCluster{
+			Properties: &ionoscloud.KubernetesClusterProperties{
+				Name:       &testClusterVar,
+				K8sVersion: &testClusterVar,
+				S3Buckets: &[]ionoscloud.S3Bucket{
+					{
+						Name: &testClusterVar,
+					},
+				},
+				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPrivateBoolVar,
 			},
 		},
 	}
@@ -101,6 +133,7 @@ var (
 					},
 				},
 				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPublicBoolVar,
 			},
 		},
 	}
@@ -122,6 +155,7 @@ var (
 					},
 				},
 				ApiSubnetAllowList: &[]string{testClusterVar},
+				Public:             &testClusterPublicBoolVar,
 			},
 			Metadata: &ionoscloud.DatacenterElementMetadata{
 				State: &testStateVar,
@@ -155,11 +189,12 @@ var (
 			Properties: &clusterProperties.KubernetesClusterProperties,
 		},
 	}
-	testClusterVar      = "test-cluster"
-	testClusterSliceVar = []string{"test-cluster"}
-	testClusterBoolVar  = false
-	testClusterNewVar   = "test-new-cluster"
-	testClusterErr      = errors.New("cluster test error")
+	testClusterVar            = "test-cluster"
+	testClusterSliceVar       = []string{"test-cluster"}
+	testClusterPrivateBoolVar = false
+	testClusterPublicBoolVar  = true
+	testClusterNewVar         = "test-new-cluster"
+	testClusterErr            = errors.New("cluster test error")
 )
 
 func TestK8sCmd(t *testing.T) {
@@ -355,11 +390,31 @@ func TestRunK8sClusterCreate(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForRequest), false)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
-		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterBoolVar)
-		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgGatewayIp), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTest, &testResponse, nil)
+		err := RunK8sClusterCreate(cfg)
+		assert.NoError(t, err)
+	})
+}
+
+func TestRunK8sClusterPrivateCreate(t *testing.T) {
+	var b bytes.Buffer
+	w := bufio.NewWriter(&b)
+	core.CmdConfigTest(t, w, func(cfg *core.CommandConfig, rm *core.ResourcesMocksTest) {
+		viper.Reset()
+		viper.Set(config.ArgQuiet, false)
+		viper.Set(config.ArgVerbose, true)
+		viper.Set(config.ArgOutput, config.DefaultOutputFormat)
+		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForState), false)
+		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForRequest), false)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPrivateBoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
+		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterPrivateTestPost).Return(&clusterPrivateTest, &testResponse, nil)
 		err := RunK8sClusterCreate(cfg)
 		assert.NoError(t, err)
 	})
@@ -376,6 +431,7 @@ func TestRunK8sClusterCreateWaitIdErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForRequest), false)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTest, nil, nil)
@@ -396,6 +452,7 @@ func TestRunK8sClusterCreateWaitErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetCluster(testClusterVar).Return(&clusterTestGet, nil, testClusterErr)
@@ -416,6 +473,7 @@ func TestRunK8sClusterCreateWaitState(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetCluster(testClusterVar).Return(&clusterTestGet, nil, nil)
@@ -437,6 +495,7 @@ func TestRunK8sClusterCreateWaitReqErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTestId, &testResponse, nil)
 		rm.CloudApiV6Mocks.Request.EXPECT().GetStatus(testRequestIdVar).Return(&testRequestStatus, nil, testRequestErr)
@@ -454,6 +513,7 @@ func TestRunK8sClusterCreateVersion(t *testing.T) {
 		viper.Set(config.ArgOutput, config.DefaultOutputFormat)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetVersion().Return(testClusterVar, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTest, nil, nil)
@@ -490,6 +550,7 @@ func TestRunK8sClusterCreateResponseErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTest, &testResponse, testClusterErr)
 		err := RunK8sClusterCreate(cfg)
@@ -509,6 +570,7 @@ func TestRunK8sClusterCreateErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testClusterVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgS3Bucket), testClusterVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgPublic), testClusterPublicBoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgApiSubnets), []string{testClusterVar})
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateCluster(clusterTestPost).Return(&clusterTest, nil, testClusterErr)
 		err := RunK8sClusterCreate(cfg)