+++ commands/cloudapi-v6/k8s_nodepool_test.go
@@ -42,6 +42,23 @@ var (
 			},
 		},
 	}
+	nodepoolTestPrivatePost = resources.K8sNodePoolForPost{
+		KubernetesNodePoolForPost: ionoscloud.KubernetesNodePoolForPost{
+			Properties: &ionoscloud.KubernetesNodePoolPropertiesForPost{
+				Name:             &testNodepoolVar,
+				NodeCount:        &testNodepoolIntVar,
+				DatacenterId:     &testNodepoolVar,
+				CpuFamily:        &testNodepoolVar,
+				AvailabilityZone: &testNodepoolVar,
+				RamSize:          &testNodepoolIntVar,
+				StorageSize:      &testNodepoolIntVar,
+				StorageType:      &testNodepoolVar,
+				CoresCount:       &testNodepoolIntVar,
+				GatewayIp:        &testNodepoolVar,
+				K8sVersion:       &testNodepoolVar,
+			},
+		},
+	}
 	nodepoolTest = resources.K8sNodePool{
 		KubernetesNodePool: ionoscloud.KubernetesNodePool{
 			Properties: &ionoscloud.KubernetesNodePoolProperties{
@@ -61,6 +78,23 @@ var (
 						Dhcp: &testK8sNodePoolLanBoolVar,
 					},
 				},
+			},
+		},
+	}
+	nodepoolTestPrivate = resources.K8sNodePool{
+		KubernetesNodePool: ionoscloud.KubernetesNodePool{
+			Properties: &ionoscloud.KubernetesNodePoolProperties{
+				Name:             &testNodepoolVar,
+				NodeCount:        &testNodepoolIntVar,
+				DatacenterId:     &testNodepoolVar,
+				CpuFamily:        &testNodepoolVar,
+				AvailabilityZone: &testNodepoolVar,
+				RamSize:          &testNodepoolIntVar,
+				StorageSize:      &testNodepoolIntVar,
+				StorageType:      &testNodepoolVar,
+				K8sVersion:       &testNodepoolVar,
+				GatewayIp:        &testNodepoolVar,
+				CoresCount:       &testNodepoolIntVar,
 			},
 		},
 	}
@@ -539,9 +573,57 @@ func TestRunK8sNodePoolCreate(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPost).Return(&nodepoolTest, &testResponse, nil)
 		err := RunK8sNodePoolCreate(cfg)
 		assert.NoError(t, err)
+	})
+}
+
+func TestRunK8sNodePoolPrivateCreate(t *testing.T) {
+	var b bytes.Buffer
+	w := bufio.NewWriter(&b)
+	core.CmdConfigTest(t, w, func(cfg *core.CommandConfig, rm *core.ResourcesMocksTest) {
+		viper.Reset()
+		viper.Set(config.ArgQuiet, false)
+		viper.Set(config.ArgOutput, config.DefaultOutputFormat)
+		viper.Set(config.ArgVerbose, true)
+		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForState), false)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sNodeCount), testNodepoolIntVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgCpuFamily), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgAvailabilityZone), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgRam), testNodepoolIntVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgStorageType), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgStorageSize), testNodepoolIntVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgCores), testNodepoolIntVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sClusterId), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDataCenterId), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgGatewayIp), testNodepoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(false, nil)
+		rm.CloudApiV6Mocks.K8s.EXPECT().GetVersion().Return(testNodepoolVar, &testResponse, nil)
+		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPrivatePost).Return(&nodepoolTestPrivate, &testResponse, nil)
+		err := RunK8sNodePoolCreate(cfg)
+		assert.NoError(t, err)
+	})
+}
+
+func TestRunK8sNodePoolPrivateCreateErr(t *testing.T) {
+	var b bytes.Buffer
+	w := bufio.NewWriter(&b)
+	core.CmdConfigTest(t, w, func(cfg *core.CommandConfig, rm *core.ResourcesMocksTest) {
+		viper.Reset()
+		viper.Set(config.ArgQuiet, false)
+		viper.Set(config.ArgOutput, config.DefaultOutputFormat)
+		viper.Set(config.ArgVerbose, true)
+		viper.Set(core.GetFlagName(cfg.NS, config.ArgWaitForState), false)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgName), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sClusterId), testNodepoolVar)
+		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDataCenterId), testNodepoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(false, nil)
+		// returns error for no gateway-ip set for private k8s cluster
+		err := RunK8sNodePoolCreate(cfg)
+		assert.Error(t, err)
 	})
 }
 
@@ -565,6 +647,7 @@ func TestRunK8sNodePoolCreateGetK8sVersionErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDataCenterId), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetVersion().Return(testNodepoolVar, nil, testNodepoolErr)
 		err := RunK8sNodePoolCreate(cfg)
 		assert.Error(t, err)
@@ -592,6 +675,7 @@ func TestRunK8sNodePoolCreateWait(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPost).Return(&nodepoolTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetNodePool(testNodepoolVar, testNodepoolVar).Return(&nodepoolTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetNodePool(testNodepoolVar, testNodepoolVar).Return(&nodepoolTestId, nil, nil)
@@ -621,6 +705,7 @@ func TestRunK8sNodePoolCreateWaitErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPost).Return(&nodepoolTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetNodePool(testNodepoolVar, testNodepoolVar).Return(&nodepoolTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetNodePool(testNodepoolVar, testNodepoolVar).Return(&nodepoolTestId, nil, testNodepoolErr)
@@ -650,6 +735,7 @@ func TestRunK8sNodePoolCreateWaitStateErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPost).Return(&nodepoolTestId, nil, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().GetNodePool(testNodepoolVar, testNodepoolVar).Return(&nodepoolTestId, nil, testNodepoolErr)
 		err := RunK8sNodePoolCreate(cfg)
@@ -678,6 +764,7 @@ func TestRunK8sNodePoolCreateErr(t *testing.T) {
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgK8sVersion), testNodepoolVar)
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgLanIds), []int{int(testNodepoolIntVar)})
 		viper.Set(core.GetFlagName(cfg.NS, cloudapiv6.ArgDhcp), testK8sNodePoolLanBoolVar)
+		rm.CloudApiV6Mocks.K8s.EXPECT().IsPublicCluster(testNodepoolVar).Return(true, nil)
 		rm.CloudApiV6Mocks.K8s.EXPECT().CreateNodePool(testNodepoolVar, nodepoolTestPost).Return(&nodepoolTest, nil, testNodepoolErr)
 		err := RunK8sNodePoolCreate(cfg)
 		assert.Error(t, err)